Default deserialization for defaultPlexDirectory
Added default deserialization for PlexDirectory. If the DirectoryDeserializer cannot find a proper subclass to deserialize a Directory, than it defaults to PlexDirectory. It's not of much use, but at least we do not have an exception.
The same approach has been user for PlexStream.
It should be done with PlexMetadata as well, but, so far, PlexMetadata is abstract so we should provide a default implementation of PlexMetadata.